[css-grid] Prettify the nested-grid.html perf test
Grid items have now a fixed height so that the grid could be seen when
opening the test with a browser.

This artificially improves the runs/s of the test because a fixed height is
much cheaper to compute than an auto height, but we do not plan to change it
so the test will still be sensitive to layout regressions.

Review-Url: https://codereview.chromium.org/2345963002
Cr-Original-Commit-Position: refs/heads/master@{#419154}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 556444f7441e1693d8cd4b0e23896f228cc238f6