[css-grid] Refactor intrinsic size computation code
The code which computes the intrinsic logical width of the grid container
and the rows sizes with indefinite height is exactly the same but operating
in different axis. Refactored so that we don't have to write the
initialization and tear down code twice.

Review-Url: https://codereview.chromium.org/2347613003
Cr-Original-Commit-Position: refs/heads/master@{#419158}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8ccb563f3b1ce1435f97a65a6a396559febf7e89