Reset VisualViewport position after same page navigation
This issue can be reproduced using pinch zoom in Android, ChromeOS and Mac.

After navigating to the **exact** same page, we create a new FrameView but
reuse the VisualViewport, meaning its scroll position remains unchanged even
though it was relative to an old FrameView.

In this patch, we reset the VisualViewport position in Page::didCommitLoad
since prior to here we would overwrite the old history item.

BUG=642279

Review-Url: https://codereview.chromium.org/2320303002
Cr-Original-Commit-Position: refs/heads/master@{#419167}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4521a9bb2402ebfffd709f02413ba3ec5a70de46