Invalidate the text autosizer on meta viewport settings changes
This patch is a followup to [1] where we noticed that viewport settings
were not invalidating the text autosizer. The text autosizer depends on
how the viewport is specified so we need to invalidate it on changes. A
test has been updated to show this works.

[1] https://crrev.com/1229b1414d64c314666d2fb3bf47706f2b5e947b

Review-Url: https://codereview.chromium.org/2341843002
Cr-Original-Commit-Position: refs/heads/master@{#419378}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 795773c25806d52ae98b77d2cdd27d0325319b0f