Sending an async GET request for doc.written blocked scripts. BUG=639401
While fetching the original request, ScriptLoader sets a member flag.
On receiving notifyFinished for the original request, HTMLScriptRunner
creates a new ScriptLoader instance using the original Element with
another flag set to true. It then calls ScriptLoader::prepareScript which
further calls fetchScript. prepareScript sets the defer flag as LazyLoad and
fetchScript sets the intervention info and header in the resource request.
The intervention info is used to set the priority in ResourceFetcher::requestResource.

This CL also adds a link in the dev tools console warning.

Also added internals.getResourcePriority() for
the layout test to test that the new request goes out with priority very low.

The code was tested using netlog to verify the second request is sent and has
the correct priority (net::IDLE).
One of the example sites created by jkarlin@ is:
http://www.karlin.me/test/docwrite/

Review-Url: https://codereview.chromium.org/2260303002
Cr-Original-Commit-Position: refs/heads/master@{#419476}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 57536eaf61f21cb9f6e351663ef75c3d6e09d705