Prevent LayoutTest setting -1 page scale factor.
Blink did not clamp page scale factor when it is -1 probably because it
is the initial value. This can result in a clusterfuzz crash as cc tries
to clamp the value it got from blink, and later found that the value (1)
does not match the one (-1) it already stores, cc then tries to update the
value and only to find that the layer tree has not be synced.

This CL prevents clusterfuzz from setting a -1 page scale factor.

BUG=640500

Review-Url: https://codereview.chromium.org/2350163002
Cr-Original-Commit-Position: refs/heads/master@{#419839}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2b5d011b487c215564e66393bd7dd1e71021ea95