[css-grid] Use grid area position to determine overflowing items
Grid code keeps a list of items overflowing their grid areas so that it
could apply some painting optimizations to reduce the amount of grid items
to be painted. Even if the border box is completely contained by the
item's grid area, margins in the item could make it overflow its grid area
by altering the border box position.

The problem is not that the optimization won't work, it's worse because it will
make that item not visible as it won't be painted.

BUG=628155

Review-Url: https://codereview.chromium.org/2353513003
Cr-Original-Commit-Position: refs/heads/master@{#420094}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a67debf9a9b78f60e7b10b5574ae4ceab49af31b