Stop serializing OffscreenCanvas index.
OffscreenCanvas doesn't transfer any out-of-band data (like ArrayBufferContents
or similar), so this index isn't used for anything on the receiving end.

Since this is a transfer-only object, it cannot be serialized to disk, so it's
safe to make this incompatible change to the wire format.

Review-Url: https://codereview.chromium.org/2358233002
Cr-Original-Commit-Position: refs/heads/master@{#420226}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 29bfa36d20e6d012da4ac5cf51293b7a9d526ced