Adds DevTools commands for forced viewport override to the Emulation domain. They are handled by the renderer's DevToolsEmulator.
This is a second step towards more flexible screenshots
(bit.ly/sized-screenshots): With this override, we can position a specific
(scaled) area of the page into the top-left corner of the frame. We can then
take a screenshot after resizing the frame (RWHV) in the browser to the scaled
area size. We are adding DevTools commands for the latter in a separate patch:
https://codereview.chromium.org/2226323002/

We enforce the override through the root layer transform of WebViewImpl.
As the given area coordinates are relative to the page origin and a page
scale of 1.0, we take the current scroll offset and page scale into account
and update the root transform when they change.

To ensure that all content in the area is recorded and shown, we override
the visible content area used in CompositedLayerMapping and disable clipping
to the visual viewport while the override is active.

BUG=625577
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2237433004
Cr-Original-Commit-Position: refs/heads/master@{#420314}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 23d8f263fccca2ef163a9c3f1592d3a51e3a2d50