Update TestExpectations for removed/renamed tests when updating w3c tests.
This change would make it so that when update-w3c-deps is run, lines in
TestExpectations for deleted or renamed tests are updated.

I believe that this wouldn't work for lines with directories instead of
individual tests as-is, although I think that's a relatively rare case.

BUG=649010

Review-Url: https://codereview.chromium.org/2360893002
Cr-Original-Commit-Position: refs/heads/master@{#420362}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 809c4145ce868eeb7bbbb6d65d08bc1eac16aaf6