Allows inline style in User-Agent shadow trees under CSP
When a page specifies a CSP of style-src 'self', it does not allow inline style
changes. However, the node under User-Agent shadow trees should be an exception
as developers are not allowed to modify the inline-style, but blink may need
it. So we add shadow dom elements as an exception under CSP.

BUG=648589

Review-Url: https://codereview.chromium.org/2359813002
Cr-Original-Commit-Position: refs/heads/master@{#420456}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bf99fb1ccc22a6b037045f6174d8a113d97f1ef7