Fix use-after-scope in SharedStyleFinderTest.
SharedStyleFinderTest::matchesRuleSet created a
ElementResolveContext instance and passed a reference
to SharedStyleFinder. Then the scope of the context
is ended, and finder retains a reference to an address
on stack that ends up being reused.

It was detected by 'ThinLTO Linux ToT' bot, but it
seems to be just a coincedence.

BUG=645295

Review-Url: https://codereview.chromium.org/2354333008
Cr-Original-Commit-Position: refs/heads/master@{#420879}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 25782de6522c605860d6e0fc265b164f3d567442