Implement NSAccessibilityScrollToVisibleAction on Mac
If you navigate to an element with VoiceOver, Chrome wasn't
scrolling it into view (unless it was focusable). To fix this
we mainly just needed to implement
NSAccessibilityScrollToVisibleAction.

To get this to work in PDF files too, we needed an extra fix
since the PDF node is outside of web content but can't scroll
itself, so we get the PDF node's coordinates and then tell the
web document to scroll to make that bounding box visible.

BUG=649861,54724

Review-Url: https://codereview.chromium.org/2365263003
Cr-Original-Commit-Position: refs/heads/master@{#420977}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7d7a94d999847d71bc9f0557fd4103efa5a8ed71