Switch RootLayerScrolls to use root transform and scroll properties
This is a followup to [1] to start using the root transform and scroll
properties with root layer scrolling enabled. With this patch, the
RLS/non-RLS property trees are no longer different.

A real codechange in this patch is to not create a scroll translation
for the FrameView unless it actually scrolls. This matches the logic
used for adding LayoutView scroll translation properties.

[1] https://crrev.com/a518f5340802393f8bdaa6b3d376dc5482049351

BUG=645615
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2364103003
Cr-Original-Commit-Position: refs/heads/master@{#421085}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0abd71cfec612222822a4e4a1bb0bc89e937a9ce