Ensure valid VisiblePosition input for inSameBlock
This patch postpones the creation of |visiblePos| in
|InsertParagraphSeparatorCommand::doApply|  so that |visiblePos|
does not get invalidated by the creation of |blockToInsert|.

Note: ideally, the creation and modification of an out-of-document
element (in this case, |blockToInsert|) shouldn't invalidate
VisiblePositions, but this is still hard to fix in the current
implementation. See crbug.com/650164.

BUG=648949

Review-Url: https://codereview.chromium.org/2372903003
Cr-Original-Commit-Position: refs/heads/master@{#421138}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 172391bd511c3871ff8c7bcd73fa9cb5c1d12043