Add support for touch-action: pinch-zoom.
This has been enabled behind an experimental web platform feature so that
we can test it on a few different devices before we activate it in the
wild. Likewise putting this as an experimental feature doesn't require
and Intent to Ship just yet. I would like to ship this at the same time
we ship touch action pan directions.

BUG=626086

Review-Url: https://codereview.chromium.org/2126323002
Cr-Original-Commit-Position: refs/heads/master@{#421214}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e9eadfcdacea7659dc909851f14215258589b5c