Change the linebreak behavior of minlength and maxlength attributes of TEXTAREA.
Change it in response to a specification change, and to improve compatibility with
Firefox and Edge.

Old behavior:
 CRLF, CR, or LF is counted as 2 because we limited the submission value.

New behavior:
 CRLF, CR, or LF is counted as 1 because we limit the API value.

We can use HTMLTextAreaElement::value().length() as is because it's the API value,
and other string input should be counted by computeLengthForAPIValue().

Approved intent-to-ship:
https://groups.google.com/a/chromium.org/d/msg/blink-dev/MkSfvCan7kY/PgniO8jwAAAJ

BUG=624361

Review-Url: https://codereview.chromium.org/2380473002
Cr-Original-Commit-Position: refs/heads/master@{#421445}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a15474ff7adff41a8956ab436ac63a7d5b7a090f